Fixes: Made Window class static, Time class added, Math fixes, and working on Main class

+++ GameEngine/Core/Vector2.cpp
@@ -9,7 +9,7 @@
 #include "Vector2.hpp"
 #include <math.h>
 
-Vector2::Vector2(int x, int y)
+Vector2::Vector2(float x, float y)
 {
     this->x = x;
     this->y = y;
